Separate instructions from their encoding
