+++ wasm/src/base/mod.rs
@@ -3,3 +3,4 @@ pub mod import;
 pub mod indices;
 pub mod memory;
 pub mod types;
+pub mod instructions;
